+++ R/utils.R
@@ -39,3 +39,8 @@
     stop("No such region.", call.=FALSE)
   )
 }
+
+set_names <- function (object = nm, nm) {
+  names(object) <- nm
+  object
+}