+++ NEWS.md
@@ -1,3 +1,7 @@
+0.3.0
+* buffer for download is now a parameter
+* added `get_query_results_metadata()` to enable retrieval of col types
+
 0.2.0
 * Split into two packages as per CRAN rJava-package suggested practice.